requirements.txt
streamlit==1.36.0
faiss-cpu==1.7.4
langchain==0.2.12
langchain-community==0.2.11
langchain-google-genai==1.0.8
google-generativeai==0.7.2
python-dotenv==1.0.1
pymongo